refactor: DispatchHandler 제거한 구조로 변경
