+++ src/main/kotlin/step3/carRacing/contoller/RaceController.kt
@@ -1,12 +1,13 @@
 package step3.carRacing.contoller
 
-import step3.carRacing.Car
+import step3.carRacing.model.Car
 import step3.carRacing.view.OutputView
+import kotlin.random.Random
 
 class RaceController {
     private fun tryRound(cars: List<Car>) {
         for (car in cars) {
-            car.tryRace()
+            car.tryRace(Random.nextInt(0, 10))
         }
     }
